Move files to a new linkage directory
