+++ link-grammar/linkage/analyze-linkage.c
