+++ link-grammar/linkage/analyze-linkage.h
