+++ link-grammar/linkage/score.c
