+++ link-grammar/post-process/post-process.c
@@ -20,13 +20,12 @@
 
 #include "api-structures.h"
 #include "error.h"
-#include "externs.h"
-#include "linkage.h"
+#include "linkage/linkage.h"
+#include "linkage/score.h"
 #include "pp_knowledge.h"
 #include "pp_linkset.h"
 #include "pp-structures.h"
 #include "resources.h"
-#include "score.h"
 #include "string-set.h"
 
 #define PP_MAX_DOMAINS 128
