+++ link-grammar/sat-solver/sat-encoder.cpp
@@ -30,15 +30,15 @@ extern "C" {
 #include "fast-sprintf.hpp"
 
 extern "C" {
-#include "analyze-linkage.h"      // for compute_link_names()
-#include "prepare/build-disjuncts.h" // for build_disjuncts_for_exp()
-#include "dict-common/dict-api.h" // for print_expression()
-#include "dict-common/dict-utils.h" // for free_Exp()
+#include "dict-common/dict-api.h"    // for print_expression()
+#include "dict-common/dict-utils.h"  // for free_Exp()
 #include "disjunct-utils.h"
-#include "linkage.h"
+#include "linkage/analyze-linkage.h" // for compute_link_names()
+#include "linkage/linkage.h"
+#include "linkage/score.h"           // for linkage_score()
+#include "prepare/build-disjuncts.h" // for build_disjuncts_for_exp()
 #include "post-process/post-process.h"
 #include "post-process/pp-structures.h"
-#include "score.h"               // for linkage_score()
 #include "tokenize/word-structures.h" // for Word_struct
 #include "tokenize/tok-structures.h"  // got Gword internals
 }
